Significant code added to deal with out-of-order message receipts, dropped messages, and links that close at inopportune times. A summary of the changes:
1. Every SNDP message has a sequence number. Messages must be processed in order, without
   skips.
2. Every SNDP message has an acknowledgment field, acknowledging the last sequential sequence
   number received.
3. And ACK messages were added, for situations where an SNDP message does not generate
   a response.
4. AddIntroduction and RemoveIntroduction messages were added, for when an introduction
   attempts to create a link that is already present.

Optimized events-processed; VPNs do not wake up and check retransmissions if there are no
messages waiting for acks.

Also, sundry bug fixes were made.

The next step: refactor SimVPNImpl, because much of the functionality in that class
is (or should be) part of the SND protocol. The VPN functionality should be
restricted to sending and receiving messages. But, before that, find out why
the performance of the simulation is so bad.

Saving the number of events processed as a property.