Refactor successful. Most of SimVPNImpl is now in a class named Link, which is in the node package. And SimComms has been renamed Router and moved into node. The Implementation interface has been streamlined--it only provides a DiscoveryService and a VPNFactory. The new SimVPN is a *MUCH* simpler class than the old SimVPNImpl; all it does is send messages (packets?) to the other side.
When a link goes to the HALF_CLOSED state, it checks for unacknowledged
messages. If there are none, it goes directly to CLOSED.

Nascent bug --- when a SimVPN is closed, it immediately closes the remote
VPN. It should send an event that has the same latency as a message, and
the event closes the other side.