This was a MAJOR refactoring. In summary:
The VPN tagging interface is gone. SND does not "manage" VPNs. The VPNFactory interface
is now a VPNManager. When VPNs are created or closed, the request is passed to the VPNManager,
and it is handled.

The Comms module now does routing. WrappedMessage is gone.

The SND Protocol (SNDP) is processed by three protocol modules: InitiateRequestProtocol,
ReceiveRequestProtocol and ReceiveOfferProtocol. None of the protocols directly sends
messages---they all construct SNDPMessageTransmissionProtocols to handle the sending of
messages. Every SNDP message is acknowledged individually. (Bug -- need to fix how acked
messages are identified to ensure there are no collisions.)

When an introduction will cause an existing VPN to be reused, the destination node sends
an AddIntroductionRequestMessage to the requester. If the message is acknowledged, then
the requester has confirmed that the VPN will be reused for the new introduction, and the
destination sends an IntroductionCompletedMessage to the introducer. (As opposed to sending
an IntroductionAcceptedMessage.)

This version of SND should map more cleanly to a real implementation. In previous versions,
simulation components had crept up into the SNDP logic.