Change class SNDNode to Node.
The Node now supports a finite number of introduced VPNs. When the
introduced VPN count exceeds that number, the LRU VPN is closed.

Made 1x2 props a subclass of 1x1.props.

Moved all of the property parsing into PropertyParser.java.

Fixed a bug and did some refactoring:

1. Bug fix -- the target of an introduction was adding the introduction to an already-
closed VPN instead of creating a new one, even though it was trying to create a
new one. Fixed the close semantics:  a) closing a VPN is delayed by one time unit to
ensure that already-transmitted messages make it out. b) one end of a VPN closes the
other by sending a CLOSE_VPN_EVENT. c) the VPN invokes a vpnClosed(addr) callback
when it is closed, allowing the Node to purge itself of the VPN.

2. Added debugging to assist with finding the above problem. When verbose is on, the
Node announces when it is opening/closing/adding IRs/removing IRs from a VPN.

3. Refactoring: the three protocols that comprise the introduction protocols have
been renamed to RequesterProtocol, IntroducerProtocol and TargetProtocol. And their
toString() names now include the address of the node to which they belong.