This is a "working" version of the system. It did, after many simulation runs, experience a bug; we will talk about that.
The big change was to change the pending feedbacks implementation. It was a map whose
values were TimeAndIntroductionRequest pairs, such that the maintenance thread could
identify which entries were too old and had to be pruned. This was massively expensive,
computationally. Replacing it with a generational implementation, such that the
maintenance action is simply to rotate the sets, clearin the oldest, made the simulations
over ten times faster.

Also abandoned the idea of serializing and deserializing messages on transmission.

Two bugs:

1. Two introduction sequences moving in the same direction attempt to create the same VPN.
   The second introduction tears down and replaces the VPN created by the first, because
   it received an IOException when attempting to transmit on the nascent link. Need to
   distinguish between nascent links (just use them, do not tear them down) vs. broken
  or non-existant links (create a new one).

2. Two introducion sequences moving in opposite directions make use of the same
   introduced connection. One direction closes the link at the wrong moment; the
   other direction is left VPN-less. Want to abandon associating IRs with VPNs
   as the means to keep them open, in favor of a LinkedHashMap where we can
   periodically close the LRU introduced links.