Cleaned up some code. Fixed a bug in statistics collection--we do not count good messages sent from attackers to victims, or their responses.
Oh, and the "bugs" mentioned in the previous bug report are not to be fixed.
They are rare events. The system fixes itself when they occur, and in a real
tcp-based system the endpoints would re-send their messages. In an async,
real-time system it is impossible to create a wholy consistant protocol.
At least, without massive synchronization.